Minor corrections. Change to the log in flow
