+++ app/controllers/sessions.js
@@ -6,28 +6,26 @@ const usersService = require('../services/users');
 exports.logIn = (req, res, next) => {
   logger.info(`${req.method} ${req.path} start... User authentication.`);
   const inUser = req.body;
-  let user = {};
   return usersService
     .findUserBy({ username: inUser.username })
-    .then(result => {
-      if (result) {
-        user = result;
-        return sessionsHelpers.comparePassword(inUser.password, user.password);
+    .then(user => {
+      if (user) {
+        return sessionsHelpers
+          .comparePassword(inUser.password, user.password)
+          .then(isPassword => (isPassword ? sessionsHelpers.generateToken(user) : null));
       }
-      return Promise.reject(errors.badLogInError('The email or password provided is incorrect'));
+      return null;
     })
-    .then(isPassword =>
-      isPassword
-        ? sessionsHelpers.generateToken(user)
-        : Promise.reject(errors.badLogInError('The email or password provided is incorrect'))
-    )
-    .then(token =>
-      res
-        .status(200)
-        .set({ authorization: token })
-        .send({
-          message: `User authenticated. Session expires at ${sessionsHelpers.getExpirationTime(token)}`
-        })
-    )
+    .then(token => {
+      if (token) {
+        return res
+          .status(200)
+          .set({ authorization: token })
+          .send({
+            message: `User authenticated. Session expires at ${sessionsHelpers.getExpirationTime(token)}`
+          });
+      }
+      return Promise.reject(errors.badLogInError('The username or password provided is incorrect'));
+    })
     .catch(next);
 };
