+++ config/index.js
@@ -41,7 +41,8 @@ const config = {
     session: {
       header_name: 'authorization',
       expirationTime: process.env.JWT_EXPIRATION_TIME,
-      secret: process.env.NODE_API_SESSION_SECRET
+      secret: process.env.NODE_API_SESSION_SECRET,
+      saltRounds: process.env.ENCRYPTION_SALT_ROUNDS
     }
   }
 };
